+++ blobmanager.js
@@ -1,13 +1,13 @@
 class BlobManager {
 	constructor() {
 		this.blobs = [];
-		this.initialBlobAmount = 4;
+		this.initialBlobAmount = 6;
 		this.standardBlobSize = 20;
 
 		colorMode(HSB, 255, 255, 255);
+		this.playerBlobColor = color(0, 255, 170);
 
 		this.initBlobs();
-		this.initPlayerBlob();
 
 		this.deadBlobs = new FIFOQueue();	
 	}
@@ -31,18 +31,21 @@ class BlobManager {
 		let vel = createVector(randomGaussian(0, 2), randomGaussian(0, 2));
 		let size = randomGaussian(this.standardBlobSize, 0.1);
 		let col = color(random(255), 120, 230);
-		let manualControl = false;
-		this.blobs.push(new Blob(size, col, pos, vel, manualControl));
+		let inputModule = new PerlinInput();
+		let isManual = false;
+		let blob = new Blob(size, col, pos, vel, isManual);
+		blob.setInputModule(inputModule);
+		this.blobs.push(blob);
 	}
 
 	initPlayerBlob() {
 		let pos = createVector(width/2, height/2);
 		let vel = createVector(0, 0);
 		let size = this.standardBlobSize;
-		let col = color(0, 255, 70);
-		let manualControl = true;
-		this.playerBlob = new Blob(size, col, pos, vel, manualControl);
-		this.blobs.push(this.playerBlob);	
+		let isManual = true;
+		this.playerBlob = new Blob(size, this.playerBlobColor, pos, vel, isManual);
+		this.blobs.push(this.playerBlob);
+		return this.playerBlob;
 	}
 
 	checkForCollisions() {