+++ blobs.js
@@ -1,4 +1,7 @@
 let blobManager;
+let manualInput;
+let joystick;
+let player;
 
 function setup() {
 	let canvas = createCanvas(
@@ -8,10 +11,16 @@ function setup() {
 
 	canvas.parent('sketch-holder');
 
+	joystick = new Joystick();
+	manualInput = new ManualInput(joystick);
 	blobManager = new BlobManager();
+	player = new Player(blobManager, manualInput);
 }
 
 function draw() {
-	background(255);
+	background(50);
+	player.update();
 	blobManager.update();
+	manualInput.update();
+	joystick.run();
 }