+++ inputenum.js
@@ -1,8 +1,9 @@
 let InputEnum = {
-	UP: "Up",
-	DOWN: "Down",
-	LEFT: "Left",
-	RIGHT: "Right"
+	// UP: "Up",
+	// DOWN: "Down",
+	// LEFT: "Left",
+	// RIGHT: "Right",
+	MOVEMENT_VECTOR: "Movement Vector"
 };
 
 Object.freeze(InputEnum);